# docs/api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "GOAT AI",
    "description": "Simon Business School Strategic Intelligence API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/api/workbench/tasks": {
      "post": {
        "tags": [
          "workbench"
        ],
        "summary": "Create and enqueue a durable workbench task",
        "description": "Create a durable queued workbench task record and enqueue execution.",
        "operationId": "post_workbench_task_api_workbench_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkbenchTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchTaskAcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Workbench runtime is not available on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/sources": {
      "get": {
        "tags": [
          "workbench"
        ],
        "summary": "List declarative workbench retrieval sources",
        "description": "List the current caller-visible retrieval sources for workbench tasks.",
        "operationId": "get_workbench_sources_route_api_workbench_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchSourcesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Workbench runtime is not available on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/tasks/{task_id}": {
      "get": {
        "tags": [
          "workbench"
        ],
        "summary": "Read one durable workbench task status",
        "description": "Read the current durable state for one workbench task.",
        "operationId": "get_workbench_task_status_api_workbench_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchTaskStatusResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "description": "Workbench runtime is not available on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/workspace-outputs": {
      "get": {
        "tags": [
          "workbench"
        ],
        "summary": "List durable workspace outputs by session or project scope",
        "description": "List visible workspace outputs for one session or project restoration scope.",
        "operationId": "list_workbench_workspace_outputs_route_api_workbench_workspace_outputs_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchWorkspaceOutputsResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "description": "Workbench runtime is not available on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/workspace-outputs/{output_id}": {
      "get": {
        "tags": [
          "workbench"
        ],
        "summary": "Read one durable workspace output",
        "description": "Read the latest durable state for one workspace output.",
        "operationId": "get_workbench_workspace_output_route_api_workbench_workspace_outputs__output_id__get",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Output Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchWorkspaceOutputPayload"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "description": "Workbench runtime is not available on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/workspace-outputs/{output_id}/exports": {
      "post": {
        "tags": [
          "workbench"
        ],
        "summary": "Export one durable workspace output into a downloadable artifact",
        "description": "Create one downloadable artifact from a durable workspace output.",
        "operationId": "export_workbench_workspace_output_route_api_workbench_workspace_outputs__output_id__exports_post",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Output Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkbenchWorkspaceOutputExportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatArtifact"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "description": "Workbench runtime is not available on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "workbench"
        ],
        "summary": "Read one durable workbench task event timeline",
        "description": "Read the current durable event timeline for one workbench task.",
        "operationId": "get_workbench_task_event_timeline_api_workbench_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchTaskEventsResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "description": "Workbench runtime is not available on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Detail",
            "description": "Human-readable message or structured validation errors."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Echo of X-Request-ID when the server generated or accepted one."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "Standard JSON error envelope (Phase 13 §13.0)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WorkbenchSourcesResponse": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/WorkbenchSourcePayload"
            },
            "type": "array",
            "title": "Sources",
            "description": "Sources the current caller can see for workbench retrieval tasks."
          }
        },
        "type": "object",
        "title": "WorkbenchSourcesResponse",
        "description": "Visible retrieval source inventory for the current caller."
      },
      "WorkbenchTaskAcceptedResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "Stable durable task identifier."
          },
          "task_kind": {
            "type": "string",
            "enum": [
              "plan",
              "browse",
              "deep_research",
              "canvas"
            ],
            "title": "Task Kind",
            "description": "Task surface requested by the client."
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "description": "Current durable task lifecycle state."
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "UTC ISO-8601 creation timestamp."
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "task_kind",
          "status",
          "created_at"
        ],
        "title": "WorkbenchTaskAcceptedResponse",
        "description": "Minimal accepted-task payload returned when a workbench task is created."
      },
...
      "WorkbenchTaskRequest": {
        "properties": {
          "task_kind": {
            "type": "string",
            "enum": [
              "plan",
              "browse",
              "deep_research",
              "canvas"
            ],
            "title": "Task Kind",
            "description": "Planned task surface requested by the client."
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt",
            "description": "Primary user instruction."
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional chat/session affinity for future task replay."
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id",
            "description": "Optional project/memory scope for future project workspaces."
          },
          "knowledge_document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Document Ids",
            "description": "Optional attached knowledge documents for retrieval-aware tasks."
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids",
            "description": "Optional retrieval source ids for browse, research, or plan context."
          },
          "connector_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Connector Ids",
            "description": "Deprecated alias for source_ids. Prefer source_ids for new clients."
          }
        },
        "type": "object",
        "required": [
          "task_kind",
          "prompt"
        ],
        "title": "WorkbenchTaskRequest",
        "description": "Request body for durable long-running workbench tasks."
      },
...
      "WorkbenchTaskStatusResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "Stable durable task identifier."
          },
          "task_kind": {
            "type": "string",
            "enum": [
              "plan",
              "browse",
              "deep_research",
              "canvas"
            ],
            "title": "Task Kind",
            "description": "Task surface requested by the client."
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "description": "Current durable task lifecycle state."
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "UTC ISO-8601 creation timestamp."
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "UTC ISO-8601 last-update timestamp."
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail",
            "description": "Optional stable human-readable failure detail."
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkbenchTaskResultPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional completed task result. Current runtime guarantees markdown for plan tasks and may include citations for browse or deep_research tasks when retrieval succeeds."
          },
          "workspace_outputs": {
            "items": {
              "$ref": "#/components/schemas/WorkbenchWorkspaceOutputPayload"
            },
            "type": "array",
            "title": "Workspace Outputs",
            "description": "Optional durable workspace outputs linked to this task. Canvas documents are the first output kind in this runtime slice."
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "task_kind",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "WorkbenchTaskStatusResponse",
        "description": "Status payload returned when polling a durable workbench task."
      },
...
    }
  },
...
}